refactor: rename strategy
