+++ src/strategies/seedify-cumulative-voting-power-hodl-staking-farming/index.ts
